Clear up some issues with fqdm vs. hostname
When we're using openstack inventory, inventory_hostname is the uuid of
the openstack server, whereas ansible_fqdn is the actual fqdn of the
host. So for things where it's important that the variable resolve to
the hostname (like looking up puppet reports on disk, we need to use
nsible_fqdn). For things like looking up hostvars in the hostvars dict,
we nede to continue using inventory_hostname.

Further, when we _delegate_to_ like we do for the post-to-puppetdb step,
ansible_fqdn is the name of the host delegated to - so it's not needed
to pass in the whoami value.

Change-Id: I420112668a016988ddb397e798f6ba983291d2f8