Consume namespaced default values
We don't necessarily want to pass in values for every role invocation,
which makes using group_vars useful. However, we don't want to force
someone to define a puppet dict in their variables if all they want to
do is pass in some values. So make default variables which read from
complex data but also allow to pass in simple values.

Depends-On: I07a143f45f2eb100c231ea1b7dd617b40f8f231c
Change-Id: Ia83210d17e97707475fd0e31dd85c88a0fcce2df