Don't hardcode puppet-3-specific config paths
This patch ensures that hiera data and puppet modules, and puppet config
are copied to the right directory depending on the current puppet
version. Since it's possible for the ansible management server and the
managed nodes to have different puppet versions, we need to account for
the possibility that the source and destination paths might be
different. We also don't need to hardcode the various config paths in
config since we're using the defaults and hardcoding them would make
them incorrect for one or the other puppet versions.

Change-Id: I164f91f9a7942e8c5f059652634ec1078ae41aae